Conflicts:
	core/src/main/java/org/fao/geonet/kernel/AccessManager.java
	core/src/main/java/org/fao/geonet/kernel/search/LuceneSearcher.java
	web-ui/src/main/resources/catalog/components/catalog/CatalogService.js
	web-ui/src/main/resources/catalog/js/custom/geocat/config.js
	web-ui/src/main/resources/catalog/js/custom/sextant/config.js
	web-ui/src/main/resources/catalog/js/custom/sextant/sextant.js
	web-ui/src/main/resources/catalog/locales/en-search.json
	web/src/main/webapp/WEB-INF/config.xml

+++ common/src/main/java/org/fao/geonet/utils/IO.java
@@ -157,7 +157,7 @@ public final class IO {
             } else {
                 Files.walkFileTree(from, new CopyAcceptedFiles(from, actualTo, filter));
             }
-        } else {
+        } else if (Files.exists(from)) {
             if (filter == null || filter.accept(from)) {
                 final Path parent = actualTo.getParent();
                 if (parent != null && !Files.exists(parent)) {
@@ -180,11 +180,13 @@ public final class IO {
             actualTo = to;
         }
 
-        final Path parent = actualTo.getParent();
-        if (!Files.exists(parent)) {
-            Files.createDirectories(parent);
+        if (Files.exists(from)) {
+            final Path parent = actualTo.getParent();
+            if (!Files.exists(parent)) {
+                Files.createDirectories(parent);
+            }
+            Files.move(from, actualTo);
         }
-        Files.move(from, actualTo);
     }
     public static boolean isEmptyDir(Path dir) throws IOException {
         try (DirectoryStream<Path> children = Files.newDirectoryStream(dir)) {
